test: MET-2188 update test case
