+++ playwright/pages/epochs-dashboard.page.ts
@@ -41,25 +41,6 @@ export function epochsDashboardPage(page: Page) {
   const viewDetailButtonWidget = page.getByTestId("epoch.detailViewEpoch.viewDetail");
   const blockTab = page.getByTestId("epoch.detailViewEpoch.blockLink");
 
-  // Epoch Detail
-  const firstBlockInEpochDetail = page.getByTestId("epochList.blockValue#0");
-  const firstBlockIdInEpochDetail = page.getByTestId("epochList.blockIdValue#0");
-  const detailTitle = page.getByTestId("detail-header-title");
-  const transationTableTitle = page.getByTestId("header.table.transactions");
-  const startTimeEpochOverview = page.getByTestId("epoch.overview.startTimeValue");
-  const endTimeEpochOverview = page.getByTestId("epoch.overview.endTimeValue");
-  const totalOutputEpochOverview = page.getByTestId("epoch.overview.totalOutputValue");
-  const blockTableTitleEpochOverview = page.getByTestId("epoch.blockList.blocksTitle");
-  const blockColumnBlockTable = page.getByTestId("epochList.blockTitle");
-  const blockIdColumnBlockTable = page.getByTestId("epochList.blockIdTitle");
-  const epochTitleColumnBlockTable = page.getByTestId("epochList.epochTitle");
-  const slotTitleColumnBlockTable = page.getByTestId("epochList.slotTitle");
-  const slotNoTitleColumnBlockTable = page.getByTestId("epochList.slotNoTitle");
-  const createdAtTitleColumnBlockTable = page.getByTestId("epochList.createdAtTitle");
-  const txCountTitleColumnBlockTable = page.getByTestId("epochList.txCountTitle");
-  const feesTitleColumnBlockTable = page.getByTestId("epochList.feesTitle");
-  const outSumTitleColumnBlockTable = page.getByTestId("epochList.outSumTitle");
-
   const openLastActiveEpochDetails = async () => {
     await expect(epochsTable).toBeVisible();
     await lastActiveEpoch.click();
@@ -105,14 +86,6 @@ export function epochsDashboardPage(page: Page) {
 
   const goToEpochsDetailFromWidgetByBlockTab = async () => {
     await blockTab.click();
-  };
-
-  const goToBlockDetailFromEpoch = async () => {
-    await firstBlockInEpochDetail.click();
-  };
-
-  const goToBlockDetailFromEpochByBlockId = async () => {
-    await firstBlockIdInEpochDetail.click();
   };
 
   const checkCurrentEpoch = async ({ lastEpoch }: { lastEpoch?: BlockfrostEpochInformationDto }) => {
@@ -204,43 +177,6 @@ export function epochsDashboardPage(page: Page) {
     await expect(viewDetailButtonWidget, "View details button on epoch widget").toHaveText("View details");
   };
 
-  const checkEpochDetailPage = async ({ currentEpoch }: { currentEpoch?: BlockfrostEpochInformationDto }) => {
-    await expect(detailTitle, "Epoch detail title").toHaveText("Epoch Details");
-
-    expect(
-      moment((await startTimeEpochOverview.textContent())?.replace(",", "")).unix(),
-      "Start time on epoch detail to equal start time epoch Blockfrost"
-    ).toEqual(currentEpoch?.start_time);
-
-    expect(
-      moment((await endTimeEpochOverview.textContent())?.replace(",", "")).unix(),
-      "End time on epoch detail to equal end time epoch Blockfrost"
-    ).toEqual(currentEpoch?.end_time);
-
-    expect(
-      +((await totalOutputEpochOverview.textContent())?.replaceAll(",", "") || 0) * 10 ** 6,
-      "Total output in epoch detail to equal total block in epoch on Blockfrost "
-    ).toEqual(+(currentEpoch?.output || 0));
-
-    await expect(blockTableTitleEpochOverview, "Block table title").toHaveText("Blocks");
-
-    await expect(blockColumnBlockTable, "Check title on finished blocks table").toHaveText("Block");
-    await expect(blockIdColumnBlockTable, "Check title on finished blocks table").toHaveText("Block ID");
-    await expect(epochTitleColumnBlockTable, "Check title on finished blocks table").toHaveText("Epoch");
-    await expect(slotTitleColumnBlockTable, "Check title on finished blocks table").toHaveText("Slot");
-    await expect(slotNoTitleColumnBlockTable, "Check title on finished blocks table").toHaveText("Absolute Slot");
-    await expect(createdAtTitleColumnBlockTable, "Check title on finished blocks table").toHaveText("Created At");
-    await expect(txCountTitleColumnBlockTable, "Check title on finished blocks table").toHaveText("Transactions");
-    await expect(feesTitleColumnBlockTable, "Check title on finished blocks table").toHaveText("Fees");
-    await expect(outSumTitleColumnBlockTable, "Check title on finished blocks table").toHaveText("Output");
-  };
-
-  const checkBlockDetailPage = async () => {
-    expect(page.url().includes("/block")).toBe(true);
-    await expect(detailTitle).toHaveText("Block Details");
-    await expect(transationTableTitle).toHaveText("Transactions");
-  };
-
   return {
     openLastActiveEpochDetails,
     openLastFinishedEpochDetails,
@@ -249,14 +185,10 @@ export function epochsDashboardPage(page: Page) {
     checkTableFinishedEpochTable,
     checkCurrentEpochWidget,
     checkFinishedEpochWidget,
-    checkEpochDetailPage,
-    checkBlockDetailPage,
     goToDashboard,
     goToEpochs,
     goToEpochsFromSidebar,
     goToEpochsDetailFromWidget,
-    goToBlockDetailFromEpoch,
-    goToBlockDetailFromEpochByBlockId,
     goToEpochsDetailFromWidgetByBlockTab,
     openWidgetCurrentEpoch,
     openWidgetFinishedEpoch,
